backend/requirements.txt
flask
flask-cors
pandas
numpy
openai
langchain
pyyaml
python-dotenv